revert other incidental changes

+++ src/libawkward/array/NumpyArray.cpp
@@ -538,7 +538,7 @@ namespace awkward {
     if (length <= 10) {
       for (int64_t i = 0;  i < length;  i++) {
         T* ptr2 = reinterpret_cast<T*>(
-            reinterpret_cast<size_t>(ptr) + (size_t)stride*((size_t)i));
+            reinterpret_cast<ssize_t>(ptr) + stride*((ssize_t)i));
         if (i != 0) {
           out << " ";
         }
@@ -559,7 +559,7 @@ namespace awkward {
     else {
       for (int64_t i = 0;  i < 5;  i++) {
         T* ptr2 = reinterpret_cast<T*>(
-            reinterpret_cast<size_t>(ptr) + (size_t)stride*((size_t)i));
+            reinterpret_cast<ssize_t>(ptr) + stride*((ssize_t)i));
         if (i != 0) {
           out << " ";
         }
@@ -579,7 +579,7 @@ namespace awkward {
       out << " ... ";
       for (int64_t i = length - 5;  i < length;  i++) {
         T* ptr2 = reinterpret_cast<T*>(
-            reinterpret_cast<size_t>(ptr) + (size_t)stride*((size_t)i));
+            reinterpret_cast<ssize_t>(ptr) + stride*((ssize_t)i));
         if (i != length - 5) {
           out << " ";
         }
